Break up google map into a separated component
